routing for user's profile ok + like incrementation

+++ src/pages/Home.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+import { Link } from "react-router-dom"
 import { useSelector, useDispatch } from 'react-redux'
 
 import { loadPosts, newPost, destroyPost, editingPost } from '../redux/actions/postsActions'
@@ -19,24 +20,23 @@ const Home = () => {
   const dispatch = useDispatch()
 
   useEffect(() => {
+    const getPosts = () => {
+      fetch("https://api-minireseausocial.mathis-dyk.fr/posts")
+        .then((response) => {
+          if (!response.ok) {
+            throw Error(response.statusText);
+          }
+          return response
+        })
+        .then((response) => response.json())
+        .then((response) => {
+          dispatch(loadPosts(response))
+        })
+    }
     if (posts.length === 0) {
       getPosts()
     }
-  }, [])
-
-  const getPosts = () => {
-    fetch("https://api-minireseausocial.mathis-dyk.fr/posts")
-      .then((response) => {
-        if (!response.ok) {
-          throw Error(response.statusText);
-        }
-        return response
-      })
-      .then((response) => response.json())
-      .then((response) => {
-        dispatch(loadPosts(response))
-      })
-  }
+  }, [posts, dispatch])
 
   const addPost = () => {
     const addPost = {
@@ -59,8 +59,8 @@ const Home = () => {
       })
       .then((response) => response.json())
       .then((response) => {
-        console.log(response)
         dispatch(newPost(response))
+        setMsg('')
       })
       .catch((error) => {
         alert(error)
@@ -87,7 +87,6 @@ const Home = () => {
       })
       .then((response) => response.json())
       .then((response) => {
-        console.log(response)
         dispatch(editingPost(response))
         setEdit("")
       })
@@ -97,12 +96,41 @@ const Home = () => {
   }
 
   const deletePost = (toDeletePost) => {
-    fetch(`https://api-minireseausocial.mathis-dyk.fr/posts/${toDeletePost.id}`, {
-      method: 'delete',
+    if (window.confirm("Le message va être supprimé")) {
+      fetch(`https://api-minireseausocial.mathis-dyk.fr/posts/${toDeletePost.id}`, {
+        method: 'delete',
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token}`
+        },
+      })
+        .then((response) => {
+          if (!response.ok) {
+            throw Error(response.statusText);
+          }
+          return response
+        })
+        .then((response) => response.json())
+        .then((response) => {
+          dispatch(destroyPost(response))
+        })
+        .catch((error) => {
+          alert(error)
+        })
+    }
+  }
+
+  const incrementLike = (post) => {
+    const data = {
+      like: post.like + 1
+    }
+    fetch(`https://api-minireseausocial.mathis-dyk.fr/posts/${post.id}`, {
+      method: 'put',
       headers: {
         'Content-Type': 'application/json',
         'Authorization': `Bearer ${token}`
       },
+      body: JSON.stringify(data)
     })
       .then((response) => {
         if (!response.ok) {
@@ -112,8 +140,8 @@ const Home = () => {
       })
       .then((response) => response.json())
       .then((response) => {
-        console.log(response)
-        dispatch(destroyPost(response))
+        dispatch(editingPost(response))
+        setEdit("")
       })
       .catch((error) => {
         alert(error)
@@ -126,30 +154,37 @@ const Home = () => {
       <p>This website is a training to Redux and React. We use auth and routing to create a small social media website.</p>
       {isAuthenticated &&
         <>
-          <h2>fuck {user.username}</h2>
+          <h2>Hello {user.username}, what do you want to share with us today ?</h2>
           <input type="text" placeholder="your shitty message" onChange={(e) => setMsg(e.target.value)} value={msg} />
           <button onClick={addPost}>Add your post</button>
         </>
       }
       <div>
         <ul>
-          {posts.map((post) => {
+          {posts.sort((a, b) => new Date(b.created_at) - new Date(a.created_at)).map((post) => {
             if (!post.text || !post.user) {
-              return
+              return false
             }
             return (
-              <li>
+              <li key={post.id}>
                 {isAuthenticated && post.user.id === user.id && edit !== post.id &&
                   <>
                     <img src={Trash} alt="delete post" onClick={() => deletePost(post)} style={{ height: "1em" }} />
                     <button onClick={() => setEdit(post.id)}>Edit</button>
                   </>
                 }
-                <b>{post.user.username}</b> :
+                {isAuthenticated &&
+                  <Link to={`/user/${post.user.id}`}>
+                    <b>{post.user.username} : </b>
+                  </Link>
+                }
                 {edit !== post.id &&
                   <>
                     {post.text}
                   </>
+                }
+                {isAuthenticated && edit !== post.id &&
+                  <span onClick={() => incrementLike(post)}> <b>|</b> {`❤ ${!post.like ? 0 : post.like}`} <b>|</b></span>
                 }
                 {edit === post.id &&
                   <>
